Uniform vs non-uniform test and graphs

+++ Code/Simple/NaiveInteger/Tests.h
@@ -108,7 +108,7 @@ inline void testSelectQuery(uint amount, uint alphabetSize, float skew, string p
     uint occurrencePack = 10;
     uint occurrenceStep = queries/occurrencePack;
     uint maxOccurrence = 1000;
-    ulong results[queries]; //just to make sure nothing is optimized away
+    int results[queries]; //just to make sure nothing is optimized away
     for(uint index = 0; index < queries; index++) {
         uint occurrence = ((index+1)*occurrenceStep) % maxOccurrence + (index+1)/occurrencePack;
         uint character = index*charStep;
